1. 외부 시그널 저장 완료 2. 외부 시그널 표시창 버그 수정완료 3. 외부 시그널 표시창 즉석 설정창 완료

+++ SMTrader/FileWatch/VtFileEventMonitor.h
@@ -143,7 +143,7 @@ public:
 			if (fs.peek() == '\n') {
 				//Start searching for \n occurrences
 				fs.seekg(-1, std::ios_base::cur);
-				int i = fs.tellg();
+				int i = (int)fs.tellg();
 				for (i; i > 0; i--) {
 					if (fs.peek() == '\n') {
 						//Found
